Got around bug on getting pellet
There was a bug where the player was immediately killed upon getting a
pellet; that was happening because sometimes the phantom ship failed to
spawn, allowing another ship to spawn in its place, sometimes in an
inconvenient position. The actual cause (phantom ship not spawned) has not
been fixed yet, but a delay was added so that enemies aren't spawned on
top of the player.